refactor: rename ha context formatter package
