+++ internal/integrations/homeassistant/contextfmt/entity_format_binary_sensor.go
@@ -1,4 +1,4 @@
-package entitycontext
+package contextfmt
 
 // binarySensorStateLabels maps a binary_sensor device_class to its
 // semantic [off, on] labels. Tracks the canonical Home Assistant
