+++ internal/integrations/homeassistant/contextfmt/entity_format_cover.go
@@ -1,4 +1,4 @@
-package entitycontext
+package contextfmt
 
 import (
 	"time"
