+++ internal/integrations/homeassistant/contextfmt/entity_format_media_player.go
@@ -1,4 +1,4 @@
-package entitycontext
+package contextfmt
 
 import (
 	"time"
